refactor(utils): 금액 추출 유틸함수들 분리
