+++ src/utils/prices/getTotalPrice.ts
